pyproject.toml
[project]
name = "stock-tracker"
version = "0.1.0"
description = "A Python CLI stock tracker with config management, API integration, and test coverage."
authors = [{ name = "Cory Gyarmathy", email = "cory.gyarmathy@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10"

# Runtime dependencies
dependencies = [
  "pytest",
  "pytest-mock",
  "yfinance",
  "pyyaml",
  "pandas",
  "pandas-stubs",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-mock", "ruff", "pyright"]

[project.scripts]
stock-tracker = "stock_tracker.main:main"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["stock_tracker"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pyright]
include = ["stock_tracker", "tests"]
pythonVersion = "3.12"
